Client and server are almost done ! The thing is almost working, it just needs more debug and testing. Added commands ! :)

+++ src/me/winter/trapgame/client/TrapButton.java
@@ -20,7 +20,7 @@ public class TrapButton extends JButton
 
 	public TrapButton(TrapGameBoard board, Point point)
 	{
-		super(point.getX() + ", " + point.getY());
+		super((int)point.getX() + ", " + (int)point.getY());
 		this.board = board;
 		this.point = point;
